Using XOrShift random numbers

+++ c++/hittable_list.hpp
@@ -84,7 +84,7 @@ public:
         Vec8f tMinVec(t_min);
         Vec8ui curId(0, 1, 2, 3, 4, 5, 6, 7);
 
-        for (int i = 0; i < radius.size(); i++)
+        for (int i = 0; i < (int)radius.size(); i++)
         {
             // load data for 4 spheres
             Vec8f ocX = rOrigX - centreX[i];
